Merge the kdev-model-and-api-experiments branch into trunk.
Nothing inherits from KDevProject anymore. This means that if your
project manager was inheriting from KDevProject like KDevProjectManager was,
then it'll need to be changed. Several methods in KDevProject have also been 
renamed. Here is the list of conversions:

openProject -> open
closeProject -> close
projectDirectory -> folder
absoluteUrl -> urlRelativeToProject
projectName -> name

KDevProject::relativeUrl has been deprecated. a name property has been
added to KDevProject with the accompanying accessors setName and name. This is
so that the name of the project no longer depends on the name of the
project file.

KDevPluginController has been given a massive overhaul. It is no longer 
managed by KDevCore. So code that used KDevCore::pluginController should
use KDevPluginController::self(). This is due, in part, to the strict nature
of the way KDevCore objects are cleaned up. 

!!! IMPORTANT !!! Plugins must have an X-KDE-PluginInfo-Name key in
their desktop file or they will not be found or loaded by the new plugin 
controller! It is preferable to fill in the whole X-KDE-PluginInfo
structure, but X-KDE-PluginInfo-Name is the only key required. 

The project model is now based on QStandardItemModel and QStandardItem.
The project model is now populated by the KDevProject object itself when
open() is called and can be accessed using the model() function.


svn path=/trunk/KDE/kdevelop/languages/java/; revision=585192