Eliminate the dependency on Flex by importing the FlexLexer.h (normally residing in /usr/include) into SVN (once for each Flex-based plugin), and also the generated *_lexer.cpp files.
As a consequence, the lexers are not anymore generated
on building, so if you need to update them, you must
run flex manually.

FlexLexer.h and *_lexer.cpp are the ones from Flex 2.5.31.
If you have a newer version lying around, say, 2.5.33,
you're welcome to get the new FlexLexer.h into SVN
and regenerate the lexer files.

svn path=/trunk/KDE/kdevelop/languages/java/; revision=590631