Camel casing in the Java parser, pt. 2: enums and flags.
Also, adapt to the new meaning for the %namespace directive:
Properly in the Java parser, semantically unchanged in the C# one.
The other parsers don't use that feature.

And no, don't ask me why kdev-pg likes to reshuffle all the
tokens in the parser's if() statements, I have no idea.
Makes for a lot of lines-changed noise.

svn path=/trunk/playground/devtools/kdevelop4-extra-plugins/java/; revision=817708