Progress in java support - almost alpha state now, just a few bugs left.
Extensive work on code completion - re-porting much code from the c++ version, which (whilst in need of cleanup + generic extraction) is still in flux.  Search for declarations as 
specified for java.
Override TopDUContext::findContextsInternal to work for java.
DeclarationBuilder:
- implement classContextOpened (needs to be moved to abstractdeclarationbuilder) to set type's declaration link
ContextBuilder:
- open context for interface and enum declarations even if no body
- rework compilation unit parsing so that imports come first, package is within a context if its qualified identifier is longer than 1 identifier (so the identifier is correct), and 
that the rest of the file is within the proper context with the package's full id
- add classContextOpened virtual function (needs to be moved to abstractcontextbuilder)


svn path=/trunk/playground/devtools/kdevelop4-extra-plugins/java/; revision=933539